EskayApp-2.0-BLL/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
click==7.1.2
distlib==0.3.1
Django==3.1.1
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
image==1.5.33
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numexpr==2.7.2
numpy==1.19.1
opencv-python
Pillow==8.0.1
pytz==2020.1
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
Werkzeug==1.0.1

